Switch from Brunch to Vite for the frontend code
