--- src/frontend/app/assets/index.html
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="utf-8">
-  <meta name="viewport" content="initial-scale=1.0" />
-  <title>Lugh</title>
-  <link rel="stylesheet" href="/app.css" type="text/css">
-  <link rel="stylesheet" href="/fonts/font-awesome.min.css" type="text/css">
-  <link rel="stylesheet" href="/flags/flag-icon.min.css" type="text/css">
-  <script src="/vendor.js"></script>
-</head>
-<body>
-  <div id="app">
-  </div>
-  <script src="/app.js"></script>
-  <script>
-    require('start.js');
-  </script>
-</body>
-</html>
