+++ src/frontend/package.json
@@ -2,31 +2,26 @@
   "name": "lugh-frontend",
   "description": "Lugh (frontend)",
   "author": "Stanislas Signoud and Raphaël Lustin",
-  "version": "0.0.1",
+  "version": "0.2.0",
   "license": "MIT",
   "repository": {
     "type": "git",
-    "url": ""
+    "url": "https://github.com/rlustin/lugh"
   },
   "scripts": {
-    "start": "brunch watch --server",
-    "prod": "brunch build --production"
+    "dev": "vite",
+    "build": "vite build",
+    "serve": "vite preview"
   },
-  "dependencies": {},
+  "dependencies": {
+    "vue": "^3.0.5",
+    "lodash": "^4.17.20"
+  },
   "devDependencies": {
-    "auto-reload-brunch": "^2.7.1",
-    "babel-brunch": "~6.1.1",
-    "babel-plugin-transform-runtime": "^6.23.0",
-    "clean-css-brunch": "^2.10.0",
-    "copycat-brunch": "github:cmelgarejo/copycat-brunch",
-    "css-brunch": "^2.10.0",
-    "flag-icon-css": "^2.8.0",
-    "font-awesome": "^4.7.0",
-    "javascript-brunch": "^2.10.0",
-    "lodash": "^4.16.6",
-    "uglify-js-brunch": "^2.10.0",
-    "vue": "^2.4.2",
-    "vue2-brunch": "^1.0.1",
-    "whatwg-fetch": "^1.0.0"
+    "@iconify/json": "^1.1.300",
+    "@vitejs/plugin-vue": "^1.1.4",
+    "@vue/compiler-sfc": "^3.0.5",
+    "vite": "^2.0.0-beta.65",
+    "vite-plugin-icons": "^0.2.1"
   }
 }
