+++ src/frontend/src/main.js
@@ -1,0 +1,13 @@
+import { createApp } from "vue";
+import App from "./App.vue";
+import Store from "./store.js";
+
+const store = new Store();
+
+window.onhashchange = function () {
+  store.namespace = window.location.hash.substring(1);
+};
+
+store.applyLocalStorage();
+
+createApp(App, { store }).mount("#app");
